ref(chat): create AbstractMessageContainer
So mobile and web can share logic for grouping chat
messages by sender.
